Re-export errors from ops and unexport ops module
This also moves errors that are specific to add to a separate file.
Otherwise rustdoc would show them to be included for the features:
"(add or not) and add", which is obviously equivalent to add but reads
super weird.
